Create Cam_Folder_Comparison.py
My dad has a folder for all the photos taken on our camera.

Within that folder are two folders (jpg files and raw files).

He wanted something that would compare the two folders and only keep the files that were in both folders.

The extra files that were in the raw folder should be moved to a new folder which would then be reviewed and then deleted.

To use this for yourself simply replace the folder paths based on your setup.