Create Groundhog_Day_Game.py
This a simple guessing game based of the movie Groundhog Day where the main character keeps waking up at 6:00am on the same day over and over.

Someone had calculated how long he was stuck in the loop for (12,395 days).

For this game the user has to guess the number of days the main character was stuck in the loop for to escape.

For each guess the user is told whether the answer is higher or lower than their input until they guess correctly.