Create Voucher_Codes.py
A simple voucher code creator that could be expanded to be used for websites.

The user inputs the discount and the number of vouchers they want to create and then they are created and displayed.

These codes could then be exported to a .txt file or a database if implemented.