Create Guest_List.py
Based of the movie Project X, 3 friends throw a birthday party and they only want 50 people. Things get out of hand as more people hear about the party.

This project adds some security which checks names against a guest list.

If the person is not on the list but there are less than 50 people total they will be allowed in otherwise they will be sent away.

Friends will be given priority as VIP guests.

For the sake of testing I limited the guest list to 5 but to increase the capacity simply increase the "guests_max" variable.